project/api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
bcrypt==3.2.0
PyJWT==2.1.0
flask-mail==0.9.1 